Fixed bugs in NearBy Section
